+++ src/test/java/dev/orne/config/AbstractConfigTest.java
@@ -37,43 +37,43 @@ class AbstractConfigTest {
 	private static final String TEST_KEY = "test.key";
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Boolean} target type.
 	 */
 	@Test
 	public void testGetBoolean() {
 		final Boolean boolValue = Boolean.TRUE;
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getBoolean(TEST_KEY)).willReturn(boolValue);
+		given(config.getBooleanParameter(TEST_KEY)).willReturn(boolValue);
 		
-		final Boolean result = config.get(TEST_KEY, Boolean.class);
+		final Boolean result = config.getParameter(TEST_KEY, Boolean.class);
 		
 		assertNotNull(result);
 		assertEquals(boolValue, result);
 		
-		then(config).should(times(1)).getBoolean(TEST_KEY);
+		then(config).should(times(1)).getBooleanParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code String} target type.
 	 */
 	@Test
 	public void testGetString() {
 		final String strValue = "test.value";
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final String result = config.get(TEST_KEY, String.class);
+		final String result = config.getParameter(TEST_KEY, String.class);
 		
 		assertNotNull(result);
 		assertEquals(strValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code URL} target type.
 	 * 
 	 * @throws MalformedURLException Not thrown
@@ -84,18 +84,18 @@ class AbstractConfigTest {
 		final String strValue = "http://projects.orne.dev/some/path";
 		final URL expectedValue = new URL(strValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final URL result = config.get(TEST_KEY, URL.class);
+		final URL result = config.getParameter(TEST_KEY, URL.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code URI} target type.
 	 */
 	@Test
@@ -103,36 +103,36 @@ class AbstractConfigTest {
 		final String strValue = "/some/path";
 		final URI expectedValue = URI.create(strValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final URI result = config.get(TEST_KEY, URI.class);
+		final URI result = config.getParameter(TEST_KEY, URI.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Character} target type.
 	 */
 	@Test
 	public void testGetChar() {
 		final String strValue = "w";
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final Character result = config.get(TEST_KEY, Character.class);
+		final Character result = config.getParameter(TEST_KEY, Character.class);
 		
 		assertNotNull(result);
 		assertEquals('w', result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Byte} target type.
 	 */
 	@Test
@@ -140,18 +140,18 @@ class AbstractConfigTest {
 		final Byte expectedValue = 123;
 		final BigDecimal numberValue = BigDecimal.valueOf(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final Byte result = config.get(TEST_KEY, Byte.class);
+		final Byte result = config.getParameter(TEST_KEY, Byte.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Short} target type.
 	 */
 	@Test
@@ -159,18 +159,18 @@ class AbstractConfigTest {
 		final Short expectedValue = 123;
 		final BigDecimal numberValue = BigDecimal.valueOf(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final Short result = config.get(TEST_KEY, Short.class);
+		final Short result = config.getParameter(TEST_KEY, Short.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Integer} target type.
 	 */
 	@Test
@@ -178,18 +178,18 @@ class AbstractConfigTest {
 		final Integer expectedValue = 123;
 		final BigDecimal numberValue = BigDecimal.valueOf(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final Integer result = config.get(TEST_KEY, Integer.class);
+		final Integer result = config.getParameter(TEST_KEY, Integer.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code BigInteger} target type.
 	 */
 	@Test
@@ -197,18 +197,18 @@ class AbstractConfigTest {
 		final BigInteger expectedValue = BigInteger.valueOf(2467894425678L);
 		final BigDecimal numberValue = new BigDecimal(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final BigInteger result = config.get(TEST_KEY, BigInteger.class);
+		final BigInteger result = config.getParameter(TEST_KEY, BigInteger.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Long} target type.
 	 */
 	@Test
@@ -216,18 +216,18 @@ class AbstractConfigTest {
 		final Long expectedValue = 123457L;
 		final BigDecimal numberValue = BigDecimal.valueOf(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final Long result = config.get(TEST_KEY, Long.class);
+		final Long result = config.getParameter(TEST_KEY, Long.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Float} target type.
 	 */
 	@Test
@@ -235,18 +235,18 @@ class AbstractConfigTest {
 		final Float expectedValue = 123.6f;
 		final BigDecimal numberValue = BigDecimal.valueOf(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final Float result = config.get(TEST_KEY, Float.class);
+		final Float result = config.getParameter(TEST_KEY, Float.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Double} target type.
 	 */
 	@Test
@@ -254,18 +254,18 @@ class AbstractConfigTest {
 		final Double expectedValue = 7534.33;
 		final BigDecimal numberValue = BigDecimal.valueOf(expectedValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final Double result = config.get(TEST_KEY, Double.class);
+		final Double result = config.getParameter(TEST_KEY, Double.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code BigDecimal} target type.
 	 */
 	@Test
@@ -273,36 +273,36 @@ class AbstractConfigTest {
 		final BigDecimal expectedValue = BigDecimal.valueOf(2467894425678.34568854);
 		final BigDecimal numberValue = expectedValue;
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getNumber(TEST_KEY)).willReturn(numberValue);
+		given(config.getNumberParameter(TEST_KEY)).willReturn(numberValue);
 		
-		final BigDecimal result = config.get(TEST_KEY, BigDecimal.class);
+		final BigDecimal result = config.getParameter(TEST_KEY, BigDecimal.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getNumber(TEST_KEY);
+		then(config).should(times(1)).getNumberParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Instant} target type.
 	 */
 	@Test
 	public void testGetInstant() {
 		final Instant expectedValue = Instant.now();
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getInstant(TEST_KEY)).willReturn(expectedValue);
+		given(config.getInstantParameter(TEST_KEY)).willReturn(expectedValue);
 		
-		final Instant result = config.get(TEST_KEY, Instant.class);
+		final Instant result = config.getParameter(TEST_KEY, Instant.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getInstant(TEST_KEY);
+		then(config).should(times(1)).getInstantParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code LocalDate} target type.
 	 */
 	@Test
@@ -310,18 +310,18 @@ class AbstractConfigTest {
 		final LocalDate expectedValue = LocalDate.now();
 		final String strValue = expectedValue.format(DateTimeFormatter.ISO_DATE);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final LocalDate result = config.get(TEST_KEY, LocalDate.class);
+		final LocalDate result = config.getParameter(TEST_KEY, LocalDate.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code LocalTime} target type.
 	 */
 	@Test
@@ -329,18 +329,18 @@ class AbstractConfigTest {
 		final LocalTime expectedValue = LocalTime.now();
 		final String strValue = expectedValue.format(DateTimeFormatter.ISO_TIME);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final LocalTime result = config.get(TEST_KEY, LocalTime.class);
+		final LocalTime result = config.getParameter(TEST_KEY, LocalTime.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code LocalDateTime} target type.
 	 */
 	@Test
@@ -348,18 +348,18 @@ class AbstractConfigTest {
 		final LocalDateTime expectedValue = LocalDateTime.now();
 		final String strValue = expectedValue.format(DateTimeFormatter.ISO_DATE_TIME);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final LocalDateTime result = config.get(TEST_KEY, LocalDateTime.class);
+		final LocalDateTime result = config.getParameter(TEST_KEY, LocalDateTime.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Date} target type.
 	 */
 	@Test
@@ -367,18 +367,18 @@ class AbstractConfigTest {
 		final Instant instantValue = Instant.now();
 		final Date expectedValue = Date.from(instantValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getInstant(TEST_KEY)).willReturn(instantValue);
+		given(config.getInstantParameter(TEST_KEY)).willReturn(instantValue);
 		
-		final Date result = config.get(TEST_KEY, Date.class);
+		final Date result = config.getParameter(TEST_KEY, Date.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getInstant(TEST_KEY);
+		then(config).should(times(1)).getInstantParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code Calendar} target type.
 	 */
 	@Test
@@ -387,18 +387,18 @@ class AbstractConfigTest {
 		final Calendar expectedValue = Calendar.getInstance();
 		expectedValue.setTimeInMillis(instantValue.toEpochMilli());
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getInstant(TEST_KEY)).willReturn(instantValue);
+		given(config.getInstantParameter(TEST_KEY)).willReturn(instantValue);
 		
-		final Calendar result = config.get(TEST_KEY, Calendar.class);
+		final Calendar result = config.getParameter(TEST_KEY, Calendar.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getInstant(TEST_KEY);
+		then(config).should(times(1)).getInstantParameter(TEST_KEY);
 	}
 
 	/**
-	 * Test method for {@link AbstractConfig#get(String, Class)} with
+	 * Test method for {@link AbstractConfig#getParameter(String, Class)} with
 	 * {@code File} target type.
 	 */
 	@Test
@@ -406,13 +406,13 @@ class AbstractConfigTest {
 		final String strValue = "/some/file/path";
 		final File expectedValue = new File(strValue);
 		final AbstractConfig config = BDDMockito.spy(AbstractConfig.class);
-		given(config.getString(TEST_KEY)).willReturn(strValue);
+		given(config.getStringParameter(TEST_KEY)).willReturn(strValue);
 		
-		final File result = config.get(TEST_KEY, File.class);
+		final File result = config.getParameter(TEST_KEY, File.class);
 		
 		assertNotNull(result);
 		assertEquals(expectedValue, result);
 		
-		then(config).should(times(1)).getString(TEST_KEY);
+		then(config).should(times(1)).getStringParameter(TEST_KEY);
 	}
 }