+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 01. Added dev.orne.config.Config
 01. Added dev.orne.config.AbstractConfig
+01. Added dev.orne.config.HierarchicalConfig